requirements.txt
appdirs==1.4.4
apturl==0.5.2
asgiref==3.3.4
astroid==2.5.6
autopep8==1.5.7
bcrypt==3.1.7
blinker==1.4
Brlapi==0.7.0
certifi==2019.11.28
chardet==3.0.4
Click==7.0
colorama==0.4.3
command-not-found==0.3
coreapi==2.3.3
coreschema==0.0.4
cryptography==2.8
cupshelpers==1.0
dbus-python==1.2.16
defer==1.0.6
distlib==0.3.1
distro==1.4.0
distro-info===0.23ubuntu1
Django==3.2.3
django-rest-swagger==2.2.0
djangorestframework==3.12.4
duplicity==0.8.12.0
entrypoints==0.3
fasteners==0.14.1
filelock==3.0.12
future==0.18.2
httplib2==0.14.0
idna==2.8
isort==5.8.0
itypes==1.2.0
Jinja2==3.0.1
keyring==18.0.1
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
lazy-object-proxy==1.6.0
libvirt-python==6.1.0
lockfile==0.12.2
louis==3.12.0
macaroonbakery==1.3.1
Mako==1.1.0
MarkupSafe==2.0.1
mccabe==0.6.1
monotonic==1.5
netifaces==0.10.4
oauthlib==3.1.0
olefile==0.46
openapi-codec==1.3.2
paramiko==2.6.0
pbr==5.5.1
pexpect==4.6.0
Pillow==7.0.0
protobuf==3.6.1
psycopg2==2.8.4
psycopg2-binary==2.8.6
pyasn1==0.4.8
pycairo==1.16.2
pycodestyle==2.7.0
pycups==1.9.73
Pygments==2.9.0
PyGObject==3.36.0
PyJWT==2.1.0
pylint==2.8.2
pymacaroons==0.13.0
PyNaCl==1.3.0
pyparsing==2.4.7
pyRFC3339==1.1
python-apt==2.0.0+ubuntu0.20.4.5
python-dateutil==2.8.1
python-debian===0.1.36ubuntu1
pytz==2018.5
pyxdg==0.26
PyYAML==5.4.1
requests==2.12.1
requests-oauthlib==0.7.0
requests-unixsocket==0.2.0
rsa==4.0
ruamel.yaml==0.17.10
ruamel.yaml.clib==0.2.4
SecretStorage==2.3.1
simplejson==3.17.2
six==1.11.0
sqlparse==0.4.1
systemd-python==234
testresources==2.0.1
toml==0.10.2
ubuntu-advantage-tools==27.2
ubuntu-drivers-common==0.0.0
ufw==0.36
unattended-upgrades==0.1
uritemplate==3.0.1
urllib3==1.25.8
usb-creator==0.3.7
vboxapi==1.0
virtualenv==20.4.2
wadllib==1.3.3
wrapt==1.12.1
xkit==0.0.0
